requirements.txt
absl-py==2.1.0
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Babel==2.13.1
beautifulsoup4==4.12.2
bleach==6.1.0
branca==0.7.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
comm==0.2.0
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distro==1.9.0
Django==5.0.1
django-appconf==1.0.6
django-cors-headers==4.3.1
django-select2==8.1.2
djangorestframework==3.14.0
docopt==0.6.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.19.0
filelock==3.13.1
fiona==1.9.6
flatbuffers==24.3.25
folium==0.16.0
fonttools==4.51.0
fqdn==1.5.1
gast==0.5.4
geopandas==0.14.4
google-pasta==0.2.0
grpcio==1.64.1
h5py==3.11.0
idna==3.4
ipykernel==6.27.0
ipython==8.17.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.4.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.1
jupyter-events==0.9.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyter_server==2.11.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.9
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.2
kaleido==0.2.1
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
lightgbm==4.3.0
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mysql==0.0.3
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.9.0
nbconvert==7.11.0
nbformat==5.9.2
nest-asyncio==1.5.8
notebook==7.0.6
notebook_shim==0.2.3
num2words==0.5.13
numpy==1.26.4
numtowords==0.0.1
opencv-python==4.10.0.82
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pdfrw==0.4
pillow==10.2.0
pipenv==2023.11.15
platformdirs==4.0.0
plotly==5.22.0
prometheus-client==0.19.0
prompt-toolkit==3.0.41
protobuf==4.25.3
psutil==5.9.6
pure-eval==0.2.2
pycparser==2.21
Pygments==2.17.1
pyparsing==3.1.2
PyPDF2==3.0.1
pyproj==3.6.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.12
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.31.0
reportlab==4.0.9
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.13.1
scikit-learn==1.4.2
scipy==1.13.0
seaborn==0.13.2
Send2Trash==1.8.2
setuptools==69.0.1
shapely==2.0.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
tabula==1.0.5
tabula-py==2.9.3
tenacity==8.3.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
termcolor==2.4.0
terminado==0.18.0
threadpoolctl==3.5.0
tinycss2==1.2.1
tornado==6.3.3
traitlets==5.13.0
types-python-dateutil==2.8.19.14
typing_extensions==4.12.2
tzdata==2023.4
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.11
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
xyzservices==2024.4.0
